requirements.txt
click
paramiko
psutil
pycrypto
pytest
